move error_code classifier to renderer analytics and slim it down
- Move error-code.ts + test from src/shared/src to
  src/renderer/src/extensions/analytics/mixpanel; drop the barrel
  re-export from shared/index.ts and repoint the three importers.
- Reuse the shared isErrorOfType identity check instead of re-deriving
  the name-vs-instanceof IPC-survival logic locally.
- Drop NODE_CODE_MAP (duplicated the fs backend's node-code knowledge
  and leaked raw codes), the native-code branch, and the acronym
  regex; a bare errno Error now passes its code through lowercased as
  a documented stopgap pending the project-wide node-error taxonomy.
- Keep the DownloadError sub-code map (documented as semantic
  unification, not casing) and class-name derivation.
