+++ src/shared/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./constants";
 export * from "./Debouncer";
 export * from "./errors";
-export * from "./error-code";
 export * from "./error-serialization";
 export * from "./download-errors";
 export * from "./types/logging";
